+++ rocketgram/api/poll.py
@@ -4,8 +4,10 @@
 
 
 from dataclasses import dataclass
+from datetime import datetime
 from typing import List, Optional
 
+from .message_entity import MessageEntity
 from .poll_option import PollOption
 from .poll_type import PollType
 
@@ -31,6 +33,10 @@ class Poll:
     poll_type: PollType
     allows_multiple_answers: bool
     correct_option_id: Optional[int]
+    explanation: Optional[str]
+    explanation_entities: Optional[List[MessageEntity]]
+    open_period: Optional[int]
+    close_date: Optional[datetime]
 
     @classmethod
     def parse(cls, data: dict) -> Optional['Poll']:
@@ -38,7 +44,11 @@ class Poll:
             return None
 
         options = [PollOption.parse(i) for i in data['options']]
+        explanation_entities = [MessageEntity.parse(d) for d in
+                                data.get('explanation_entities')] if 'explanation_entities' in data else None
+        close_date = datetime.utcfromtimestamp(data['close_date']) if 'close_date' in data else None
 
         return cls(data['id'], data['question'], options, data['total_voter_count'], data['is_closed'],
-                   data['is_anonymous'], PollType(data['poll_type']), data['allows_multiple_answers'],
-                   data['correct_option_id'])
+                   data['is_anonymous'], PollType(data['type']), data['allows_multiple_answers'],
+                   data.get('correct_option_id'), data.get('explanation'), explanation_entities,
+                   data.get('open_period'), close_date)