+++ rocketgram/api/send_dice.py
@@ -18,6 +18,7 @@ class SendDice(MessageResultMixin, Request):
     """
 
     chat_id: Union[int, str]
+    emoji: Optional[str] = None
     disable_notification: Optional[bool] = None
     reply_to_message_id: Optional[int] = None
     reply_markup: Optional[ALL_KEYBOARDS] = None